+++ src/components/dropdown/index.ts
@@ -1,2 +1,2 @@
 export {WsDropdown} from './ws-dropdown.js';
-export type {WsDropdownSize} from './ws-dropdown.js';
+export type {WsDropdownSize, WsDropdownVariant} from './ws-dropdown.js';